run it with:
```python
from macrobond_data_api.util import transfer_performance_test

transfer_performance_test()
```
And  output
```
- transfer performance test beginning -

- Running speed test -
Speed 45.12 MB/s

- Running integrity test -
Testing sizes [10 kB, 100 kB, 1 MB, 10 MB]

Testing 10 kB .... done
Testing 100 kB .... done
Testing 1 MB .... done
Testing 10 MB .... done

All done, No errors

Result for 10 kB
Average time 0.08 s
Max time 0.08 s
Min time 0.07 s

Result for 100 kB
Average time 0.09 s
Max time 0.09 s
Min time 0.07 s

Result for 1 MB
Average time 0.14 s
Max time 0.19 s
Min time 0.09 s

Result for 10 MB
Average time 0.42 s
Max time 0.50 s
Min time 0.24 s

- transfer performance test end -
```